+++ packages/web-client/tests/acceptance/create-merchant-persistence-test.ts
@@ -12,7 +12,7 @@ import {
   createMerchantSafe,
   createPrepaidCardSafe,
   createSafeToken,
-} from '../helpers/data';
+} from '../helpers/mocks';
 
 interface Context extends MirageTestContext {}
