+++ packages/web-client/tests/acceptance/create-merchant-test.ts
@@ -22,7 +22,7 @@ import { setupHubAuthenticationToken } from '../helpers/setup';
 
 import { MirageTestContext } from 'ember-cli-mirage/test-support';
 import { formatAmount } from '@cardstack/web-client/helpers/format-amount';
-import { createPrepaidCardSafe } from '../helpers/data';
+import { createPrepaidCardSafe } from '../helpers/mocks';
 
 interface Context extends MirageTestContext {}
