+++ packages/web-client/tests/integration/components/card-pay/withdrawal-workflow/choose-balance-test.ts
@@ -12,7 +12,7 @@ import {
   createMerchantSafe,
   createPrepaidCardSafe,
   createSafeToken,
-} from '@cardstack/web-client/tests/helpers/data';
+} from '@cardstack/web-client/tests/helpers/mocks';
 
 module(
   'Integration | Component | card-pay/withdrawal-workflow/choose-balance',
